1.3

+++ src/main/java/com/mcp/validate/validator/Max.java
@@ -16,8 +16,8 @@ public class Max extends BindingValidator {
         if (StringUtils.isEmpty(check.max())) {
             return true;
         }
-        BigDecimal cond = BigDecimal.valueOf(Long.valueOf(check.max()));
-        BigDecimal param = BigDecimal.valueOf(Long.valueOf(value.toString()));
+        BigDecimal cond = BigDecimal.valueOf(Double.valueOf(check.max()));
+        BigDecimal param = BigDecimal.valueOf(Double.valueOf(value.toString()));
         if (param.compareTo(cond) <= 0) {
             return true;
         }
